fix(s3s-fs): fix checksum for range (#285)
* Add failing test for `range`

This adds a (currently failing) test for get_object when using `range`.
It appears that the checksum of the full object is being returned.

* Fix checksum for range

This fixes the checksum behavior for range to line up with S3's
implementation. S3 will only return checksums if the range is large
enough that the entire file is returned, otherwise they get skipped.